Move get_children_recursive and prop_array_to_vector
